+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Barebones boot test now fails if user specifies a compute node that is not available.
 
 ## [1.4.0] - 2022-07-13
+
 ### Added
 - Enabled gitversion and gitflow